+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Current (2025-09-11)
 --------------------
 
+* Added CLI interface for hitmap plotting
 * Added a utility to generate an example config file for the hitmap generation
 * Added a new entry point for the sampiclyser utilities, everything that is not directly related to analysing the data but required for it
 * Fixed hints in the CLI